requirements.txt
Fabric
Jinja2
PyYAML
boto
python-dateutil
ipython
GitPython
apache-libcloud>=0.14.0-beta3
pptable
lxml
